# Improved efficiency in DigestManager.verify()