--- a/circe-checksum/src/main/java/com/scurrilous/circe/checksum/Java8IntHash.java
+++ b/circe-checksum/src/main/java/com/scurrilous/circe/checksum/Java8IntHash.java
@@ -36,11 +36,21 @@ public class Java8IntHash implements IntHash {
 
     @Override
     public int resume(int current, ByteBuf buffer) {
+        return resume(current, buffer, buffer.readerIndex(), buffer.readableBytes());
+    }
+
+    @Override
+    public int calculate(ByteBuf buffer, int offset, int len) {
+        return resume(0, buffer, offset, len);
+    }
+
+    @Override
+    public int resume(int current, ByteBuf buffer, int offset, int len) {
         if (buffer.hasArray()) {
-            return hash.resume(current, buffer.array(), buffer.arrayOffset() + buffer.readerIndex(),
-                    buffer.readableBytes());
+            return hash.resume(current, buffer.array(), buffer.arrayOffset() + offset,
+                    len);
         } else {
-            return hash.resume(current, buffer.nioBuffer());
+            return hash.resume(current, buffer.slice(offset, len).nioBuffer());
         }
     }
 }