### Motivation

In `DigestManager.verifyDigestAndReturnData()` there are a couple of simple improvements: 

 1. For the temp buffer where to store the digest, do not acquire a buffer from the pool each time. Instead, use a thread-local buffer.
 2. Instead of taking slices() at each time, extend the interface to take a "offset, length" pair and save the allocations.
 

### Benchmarks: 

#### Before

```
Benchmark                            (entrySize)   Mode  Cnt   Score   Error   Units
DigestManagerBenchmark.verifyDigest           64  thrpt    3  13.450 ± 3.634  ops/us
DigestManagerBenchmark.verifyDigest         1024  thrpt    3   7.908 ± 2.637  ops/us
DigestManagerBenchmark.verifyDigest         4086  thrpt    3   3.233 ± 0.882  ops/us
DigestManagerBenchmark.verifyDigest         8192  thrpt    3   1.846 ± 0.047  ops/us
```

Allocations: 230 bytes per operation

#### After

```
Benchmark                            (entrySize)   Mode  Cnt   Score   Error   Units
DigestManagerBenchmark.verifyDigest           64  thrpt    3  46.312 ± 7.414  ops/us
DigestManagerBenchmark.verifyDigest         1024  thrpt    3  13.379 ± 1.069  ops/us
DigestManagerBenchmark.verifyDigest         4086  thrpt    3   3.787 ± 0.059  ops/us
DigestManagerBenchmark.verifyDigest         8192  thrpt    3   1.956 ± 0.052  ops/us
```

Allocations: 64 bytes per operation

Note: these remaining 64 bytes are only due to the Java9 reflection access. It would not be present on the JNI-based CRC32c backend (though that does not work on Mac M1).

![image](https://user-images.githubusercontent.com/62500/221421437-702e4fa0-a2b3-41bb-aefa-bd8368fff4b9.png)